* commit 'eac548c65c30b04983a8c904f9a1c6003d428196':
  hwc test: 64-bit compilation error

+++ opengl/tests/hwc/hwcStress.cpp
@@ -574,8 +574,8 @@ void initFrames(unsigned int seed)
         // mod the wMod/hMod value must be equal to 0.
         size_t w = (width * maxSizeRatio) * testRandFract();
         size_t h = (height * maxSizeRatio) * testRandFract();
-        w = max(1u, w);
-        h = max(1u, h);
+        w = max(size_t(1u), w);
+        h = max(size_t(1u), h);
         if ((w % formatPtr->wMod) != 0) {
             w += formatPtr->wMod - (w % formatPtr->wMod);
         }
